Refactor the action commands dispatching logic into a nice mixin
