+++ lib/msf/ui/console/command_dispatcher.rb
@@ -134,6 +134,7 @@ module CommandDispatcher
 end
 end end end
 
+require 'msf/ui/console/module_action_commands'
 require 'msf/ui/console/module_command_dispatcher'
 require 'msf/ui/console/module_option_tab_completion'
 require 'msf/ui/console/command_dispatcher/core'
