+++ lib/msf/ui/console/command_dispatcher/auxiliary.rb
@@ -12,25 +12,8 @@ module CommandDispatcher
 class Auxiliary
 
   include Msf::Ui::Console::ModuleCommandDispatcher
+  include Msf::Ui::Console::ModuleActionCommands
   include Msf::Ui::Console::ModuleOptionTabCompletion
-
-  @@auxiliary_action_opts = Rex::Parser::Arguments.new(
-    '-h' => [ false, 'Help banner.'                                                        ],
-    '-j' => [ false, 'Run in the context of a job.'                                        ],
-    '-o' => [ true,  'A comma separated list of options in VAR=VAL format.'                ],
-    '-q' => [ false, 'Run the module in quiet mode with no output'                         ]
-  )
-
-  @@auxiliary_opts = Rex::Parser::Arguments.new(@@auxiliary_action_opts.fmt.merge(
-    '-a' =>  [ true,  'The action to use.  If none is specified, ACTION is used.'],
-  ))
-
-  #
-  # Returns the hash of commands specific to auxiliary modules.
-  #
-  def action_commands
-      mod.actions.map { |action| [action.name.downcase, action.description] }.to_h
-  end
 
   #
   # Returns the hash of commands specific to auxiliary modules.
@@ -44,38 +27,7 @@ class Auxiliary
       "recheck"  => "This is an alias for the rcheck command",
       "rexploit" => "This is an alias for the rerun command",
       "reload"   => "Reloads the auxiliary module"
-    }).merge( (mod ? mod.auxiliary_commands : {}) ).merge(action_commands)
-  end
-
-  #
-  # Allow modules to define their own commands
-  #
-  def method_missing(meth, *args)
-    if (mod and mod.respond_to?(meth.to_s, true) )
-
-      # Initialize user interaction
-      mod.init_ui(driver.input, driver.output)
-
-      return mod.send(meth.to_s, *args)
-    end
-
-    action = meth.to_s.delete_prefix('cmd_')
-    if mod && mod.kind_of?(Msf::Module::HasActions) && mod.actions.map(&:name).any? { |a| a.casecmp?(action) }
-       return do_action(action, *args)
-    end
-
-    return
-  end
-
-  #
-  #
-  # Execute the module with a set action
-  #
-  def do_action(meth, *args)
-    action = mod.actions.find { |action| action.name.casecmp?(meth) }
-    raise Msf::MissingActionError.new(meth) if action.nil?
-
-    cmd_run(*args, action: action.name)
+    }).merge( (mod ? mod.auxiliary_commands : {}) )
   end
 
   #
@@ -87,15 +39,6 @@ class Auxiliary
   end
 
   #
-  # Tab completion for the run command
-  #
-  def cmd_run_tabs(str, words)
-    flags = @@auxiliary_opts.fmt.keys
-    options = tab_complete_option(str, words)
-    flags + options
-  end
-
-  #
   # Executes an auxiliary module
   #
   def cmd_run(*args, action: nil)
@@ -104,7 +47,7 @@ class Auxiliary
     jobify  = false
     quiet   = false
 
-    @@auxiliary_opts.parse(args) do |opt, idx, val|
+    @@module_opts.parse(args) do |opt, idx, val|
       case opt
       when '-j'
         jobify = true
@@ -208,14 +151,7 @@ class Auxiliary
     print_line "Usage: run [options]"
     print_line
     print_line "Launches an auxiliary module."
-    print @@auxiliary_opts.usage
-  end
-
-  def cmd_action_help(action)
-    print_line "Usage: " + action.downcase + " [options]"
-    print_line
-    print_line "Launches an auxiliary module."
-    print @@auxiliary_action_opts.usage
+    print @@module_opts.usage
   end
 
   alias cmd_exploit_help cmd_run_help
